feat(DiscreteMovementRule): 공통 로직 추상 클래스로 분리
