+++ src/main/java/chess/domain/movement/discrete/BlackPawnDefaultMovement.java
@@ -1,0 +1,9 @@
+package chess.domain.movement.discrete;
+
+public class BlackPawnDefaultMovement extends DiscreteMovementRule {
+
+    @Override
+    protected boolean isMovable(int rankDifference, int fileDifference) {
+        return rankDifference == -1 && fileDifference == 0;
+    }
+}
