+++ src/main/java/chess/domain/movement/discrete/KingMovement.java
@@ -1,0 +1,12 @@
+package chess.domain.movement.discrete;
+
+public class KingMovement extends DiscreteMovementRule {
+
+    @Override
+    protected boolean isMovable(int rankDifference, int fileDifference) {
+        int rankDifferenceSize = Math.abs(rankDifference);
+        int fileDifferenceSize = Math.abs(fileDifference);
+
+        return rankDifferenceSize <= 1 && fileDifferenceSize <= 1;
+    }
+}
