+++ src/main/java/chess/domain/movement/discrete/KnightMovement.java
@@ -1,0 +1,13 @@
+package chess.domain.movement.discrete;
+
+public class KnightMovement extends DiscreteMovementRule {
+
+    @Override
+    protected boolean isMovable(int rankDifference, int fileDifference) {
+        int rankDifferenceSize = Math.abs(rankDifference);
+        int fileDifferenceSize = Math.abs(fileDifference);
+
+        return (rankDifferenceSize == 2 && fileDifferenceSize == 1)
+                || (fileDifferenceSize == 2 && rankDifferenceSize == 1);
+    }
+}
