+++ src/main/java/chess/domain/movement/discrete/WhitePawnDefaultMovement.java
@@ -1,0 +1,9 @@
+package chess.domain.movement.discrete;
+
+public class WhitePawnDefaultMovement extends DiscreteMovementRule {
+
+    @Override
+    protected boolean isMovable(int rankDifference, int fileDifference) {
+        return rankDifference == 1 && fileDifference == 0;
+    }
+}
