+++ src/test/java/chess/domain/movement/discrete/BlackPawnDefaultMovementTest.java
@@ -1,4 +1,4 @@
-package chess.domain.movement;
+package chess.domain.movement.discrete;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -10,39 +10,37 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-class WhitePawnDefaultMovementTest {
-
+class BlackPawnDefaultMovementTest {
 
     @Test
     @DisplayName("이동 가능한지 확인한다.")
     void isMovableTest() {
         Position start = new Position(File.D, Rank.FOUR);
-        Position end = new Position(File.D, Rank.FIVE);
-        WhitePawnDefaultMovement whitePawnDefaultMovement = new WhitePawnDefaultMovement();
+        Position end = new Position(File.D, Rank.THREE);
+        BlackPawnDefaultMovement blackPawnDefaultMovement = new BlackPawnDefaultMovement();
 
-        assertThat(whitePawnDefaultMovement.isMovable(start, end)).isTrue();
+        assertThat(blackPawnDefaultMovement.isMovable(start, end)).isTrue();
     }
 
     @ParameterizedTest
-    @CsvSource({"B, SIX", "F, SIX", "D, THREE", "D, SIX"})
+    @CsvSource({"B, SIX", "F, SIX", "D, TWO", "D, FIVE"})
     @DisplayName("이동 가능한지 확인한다.")
     void isMovableTest_false(File file, Rank rank) {
         Position start = new Position(File.D, Rank.FOUR);
         Position end = new Position(file, rank);
-        WhitePawnDefaultMovement whitePawnDefaultMovement = new WhitePawnDefaultMovement();
+        BlackPawnDefaultMovement blackPawnDefaultMovement = new BlackPawnDefaultMovement();
 
-        assertThat(whitePawnDefaultMovement.isMovable(start, end)).isFalse();
+        assertThat(blackPawnDefaultMovement.isMovable(start, end)).isFalse();
     }
 
     @Test
     @DisplayName("이동 경로를 알 수 있다.")
     void findPathTest() {
         Position start = new Position(File.D, Rank.FOUR);
-        Position end = new Position(File.D, Rank.FIVE);
-        WhitePawnDefaultMovement whitePawnDefaultMovement = new WhitePawnDefaultMovement();
+        Position end = new Position(File.D, Rank.THREE);
+        BlackPawnDefaultMovement blackPawnDefaultMovement = new BlackPawnDefaultMovement();
 
-        assertThat(whitePawnDefaultMovement.findPath(start, end))
+        assertThat(blackPawnDefaultMovement.findPath(start, end))
                 .containsExactly(end);
     }
-
 }
