+++ src/test/java/chess/domain/movement/discrete/KingMovementTest.java
@@ -1,4 +1,4 @@
-package chess.domain.movement;
+package chess.domain.movement.discrete;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -10,37 +10,38 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-class BlackPawnDefaultMovementTest {
+class KingMovementTest {
 
-    @Test
+    @ParameterizedTest
+    @CsvSource({"D, FIVE", "E, FIVE", "E, FOUR", "E, THREE", "D, THREE", "C, THREE", "C, FOUR", "C, FIVE"})
     @DisplayName("이동 가능한지 확인한다.")
-    void isMovableTest() {
+    void isMovableTest(File file, Rank rank) {
         Position start = new Position(File.D, Rank.FOUR);
-        Position end = new Position(File.D, Rank.THREE);
-        BlackPawnDefaultMovement blackPawnDefaultMovement = new BlackPawnDefaultMovement();
+        Position end = new Position(file, rank);
+        KingMovement kingMovement = new KingMovement();
 
-        assertThat(blackPawnDefaultMovement.isMovable(start, end)).isTrue();
+        assertThat(kingMovement.isMovable(start, end)).isTrue();
     }
 
     @ParameterizedTest
-    @CsvSource({"B, SIX", "F, SIX", "D, TWO", "D, FIVE"})
+    @CsvSource({"B, SIX", "F, SIX", "B, TWO", "F, TWO"})
     @DisplayName("이동 가능한지 확인한다.")
     void isMovableTest_false(File file, Rank rank) {
         Position start = new Position(File.D, Rank.FOUR);
         Position end = new Position(file, rank);
-        BlackPawnDefaultMovement blackPawnDefaultMovement = new BlackPawnDefaultMovement();
+        KingMovement kingMovement = new KingMovement();
 
-        assertThat(blackPawnDefaultMovement.isMovable(start, end)).isFalse();
+        assertThat(kingMovement.isMovable(start, end)).isFalse();
     }
 
     @Test
     @DisplayName("이동 경로를 알 수 있다.")
     void findPathTest() {
         Position start = new Position(File.D, Rank.FOUR);
-        Position end = new Position(File.D, Rank.THREE);
-        BlackPawnDefaultMovement blackPawnDefaultMovement = new BlackPawnDefaultMovement();
+        Position end = new Position(File.D, Rank.FIVE);
+        KingMovement kingMovement = new KingMovement();
 
-        assertThat(blackPawnDefaultMovement.findPath(start, end))
+        assertThat(kingMovement.findPath(start, end))
                 .containsExactly(end);
     }
 }
